Refactor: split FixpSession inbound surface into partial file (#121, step 4/4)
Closes #121. Final step of the FixpSession refactor.

After PRs #126 and #127 extracted FixpOutboundEncoder and
FixpRetransmitController as standalone collaborators, the remaining
~1000 lines of inbound surface (RunReceiveLoopAsync, DispatchInboundAsync,
ProcessNegotiate / ExecuteNegotiateStepAsync, ProcessEstablish /
ExecuteEstablishStepAsync, business-header validators, throttle gate,
WriteApplicationBusinessReject, MapHeaderErrorToTerminationCode,
RunWatchdogLoopAsync, ReadExactlyAsync helpers, EnqueueSequence
delegate) are tightly coupled to FixpSession's private state machine
fields (_msgSeqNum, _isOpen, _isAttached, _cts, _attachLock, _claims,
_validator, _establishValidator, _throttle, _suspendedSinceMs,
_recvTask, _watchdogTask, _probeOutstanding, _lastInbound* timestamps).

Extracting them as a separate collaborator class would either require
exposing all of those internals (break encapsulation) or pass 20+
ctor delegates (worse than the original). Using C# partial-class
file-organization instead is the right tool here:

- FixpSession.cs (880 lines) — class identity, ctor, public properties,
  state machine (FixpState, ApplyTransition), Close/Suspend/re-attach,
  CoD timer, write delegates to OutboundEncoder.
- FixpSession.Inbound.cs (1041 lines) — receive loop, dispatch switch,
  Negotiate/Establish handshakes, business-header validation, throttle,
  watchdog. All marked partial; no API change.

This grouping mirrors how the file already implicitly split between
"session lifecycle" and "inbound dispatch" concerns; making the split
physical removes the mental tax of scrolling between unrelated sections
and lets future Negotiate/Establish work touch one file instead of one
massive file.

The original goal (FixpSession.cs <600 lines) is not achieved literally
— a clean target would have required exposing the dozen+ internals
listed above. Pragmatically the file is now 880 lines of cohesive
lifecycle code, with the inbound surface in its own 1041-line companion;
together with the two extracted collaborators (FixpOutboundEncoder,
FixpRetransmitController) the original 2159-line monolith is now
spread across 4 focused files.

Cumulative for #121:
- FixpSession.cs:           2159 -> 880 lines (-1279)
- FixpSession.Inbound.cs:   +1041 lines (new, was inline)
- FixpOutboundEncoder.cs:    +251 lines (PR #126)
- FixpRetransmitController.cs: +214 lines (PR #127)

Verification: 595/595 tests pass (376/376 gateway), format clean.

Closes #121.
